+++ src/components/charts/bars/BarChartStackedHorizontalPro/definition.ts
@@ -10,6 +10,8 @@ import {
   loadDataResultsAxisOrder,
   loadDataResults,
 } from '../bars.loadData.utils';
+import { getClientContextTimezone } from '../../../../theme/utils/clientContext.utils';
+import { ThemeClientContext } from '../../../../theme/theme.types';
 
 const meta = {
   name: 'BarChartStackedHorizontalPro',
@@ -86,9 +88,11 @@ const props = (
     BarChartStackedHorizontalProState,
     (state: BarChartStackedHorizontalProState) => void,
   ],
+  clientContext: ThemeClientContext,
 ) => {
   const yAxisWithGranularity = getDimensionWithGranularity(inputs.yAxis, state?.granularity);
   const sortDirection = inputs.sortDirectionTopYAxis as OrderDirection | undefined;
+  const timezone = getClientContextTimezone(clientContext?.timezone);
 
   const axisOrderCacheKey = getAxisOrderCacheKey({
     dataset: inputs.dataset,
@@ -96,6 +100,7 @@ const props = (
     measure: inputs.measure,
     sortDirection,
     limit: inputs.limitTopYAxis,
+    timezone,
   });
 
   const cachedAxisOrder = getCachedAxisOrder(axisOrderCacheKey, state);
@@ -114,6 +119,7 @@ const props = (
       axis: yAxisWithGranularity,
       measure: inputs.measure,
       sortDirection,
+      timezone,
     }),
     results: loadDataResults({
       dataset: inputs.dataset,
@@ -124,6 +130,7 @@ const props = (
       limitTopAxis: inputs.limitTopYAxis,
       maxResults: inputs.maxResults,
       axisOrder: cachedAxisOrder,
+      timezone,
     }),
   };
 };