+++ src/components/charts/bars/BarChartStackedPro/definition.ts
@@ -10,6 +10,8 @@ import {
   loadDataResultsAxisOrder,
   loadDataResults,
 } from '../bars.loadData.utils';
+import { getClientContextTimezone } from '../../../../theme/utils/clientContext.utils';
+import { ThemeClientContext } from '../../../../theme/theme.types';
 
 const meta = {
   name: 'BarChartStackedPro',
@@ -83,9 +85,11 @@ const events = {
 const props = (
   inputs: Inputs<typeof meta>,
   [state, setState]: [BarChartStackedProState, (state: BarChartStackedProState) => void],
+  clientContext: ThemeClientContext,
 ) => {
   const xAxisWithGranularity = getDimensionWithGranularity(inputs.xAxis, state?.granularity);
   const sortDirection = inputs.sortDirectionTopXAxis as OrderDirection | undefined;
+  const timezone = getClientContextTimezone(clientContext?.timezone);
 
   const axisOrderCacheKey = getAxisOrderCacheKey({
     dataset: inputs.dataset,
@@ -93,6 +97,7 @@ const props = (
     measure: inputs.measure,
     sortDirection,
     limit: inputs.limitTopXAxis,
+    timezone,
   });
 
   const cachedAxisOrder = getCachedAxisOrder(axisOrderCacheKey, state);
@@ -111,6 +116,7 @@ const props = (
       axis: xAxisWithGranularity,
       measure: inputs.measure,
       sortDirection,
+      timezone,
     }),
     results: loadDataResults({
       dataset: inputs.dataset,
@@ -121,6 +127,7 @@ const props = (
       limitTopAxis: inputs.limitTopXAxis,
       maxResults: inputs.maxResults,
       axisOrder: cachedAxisOrder,
+      timezone,
     }),
   };
 };